added methods for admin

+++ public/script/scriptContent.js
@@ -114,31 +114,34 @@ function refreshContent(users) {
     var cardsContainer = document.getElementById('content');
     cardsContainer.innerHTML = '';
 
+    const loggedInClient = clientManager.client;
+
     users.forEach(user => {
-        if(!user.isDeleted){
+        if (!user.isDeleted) {
             var banBtn = '';
             var cardState = '';
             var skills = user.skillset;
             var skillsHtmlString = '';
-    
+
             skills.forEach(skill => {
                 skillsHtmlString += '<a class="skillSpan"><li class="skill">' + skill.skillName + '</li><span class="skillLevel">' + skill.skillRate + '/10</span></a>';
             });
-    
-            var onBanClick = 'javascript:requestBan("' +  user.email + '")';
-    
-            if (user.isBanned) {
-                banBtn = `<a href='`+ onBanClick + `' class='adminBtn unbanBtn'><span class='deleteSpan unbanSpan'>UNBAN</span><i class='fas fa-user-plus'></i></a>`;
-                cardState = 'bannedCard';
-            } else {
-            
-                banBtn = `<a href='` + onBanClick + `' class='adminBtn banBtn'><span class='deleteSpan banSpan'>BAN</span><i class='fas fa-user-minus'></i></a>`;
-                cardState = 'notBannedCard';
+
+            var onBanClick = 'javascript:requestBan("' + user.email + '")';
+
+            if (loggedInClient.isAdmin) {
+                if (user.isBanned) {
+                    banBtn = `<a href='` + onBanClick + `' class='adminBtn unbanBtn'><span class='deleteSpan unbanSpan'>UNBAN</span><i class='fas fa-user-plus'></i></a>`;
+                    cardState = 'bannedCard';
+                } else {
+                    banBtn = `<a href='` + onBanClick + `' class='adminBtn banBtn'><span class='deleteSpan banSpan'>BAN</span><i class='fas fa-user-minus'></i></a>`;
+                    cardState = 'notBannedCard';
+                }
+
+                var onDeleteClick = 'javascript:requestDelete("' + user.email + '")';
+                deleteBtn = `<a href='` + onDeleteClick + `' class='adminBtn banBtn'><span class='deleteSpan banSpan'>DELETE</span><i class="fas fa-trash-alt"></i></a>`;
             }
-    
-            var onDeleteClick = 'javascript:requestDelete("' + user.email + '")';
-            deleteBtn = `<a href='` + onDeleteClick + `' class='adminBtn banBtn'><span class='deleteSpan banSpan'>DELETE</span><i class="fas fa-trash-alt"></i></a>`;
-    
+
             cardsContainer.insertAdjacentHTML('beforeend', '<div id="' + user.email + '" class="card ' + cardState + '">\
                     <div class="cardHeader"><div class="adminContent">' + deleteBtn + banBtn + '</div>' +
                 '<h1 class="title">' + user.name + '</h1>\
@@ -166,28 +169,20 @@ function refreshContent(users) {
 }
 
 function requestBan(email) {
-    var  isBanned = document.getElementById(email);
+    var isBanned = document.getElementById(email);
 
-    if(isBanned.classList.contains('bannedCard')){
+    if (isBanned.classList.contains('bannedCard')) {
         //Card is banned, when clicked means it should get unbanned.
-        apiRequest.setBan(false,email);
+        apiRequest.setBan(false, email);
     } else {
         //Card is not banned, when clicked means it should get banned.
-        apiRequest.setBan(true,email);
+        apiRequest.setBan(true, email);
     }
 }
 
 function requestDelete(email) {
     //No true/false when deleting, one should only be able to delete the user and not add back him from the GUI.
     apiRequest.setDelete(email);
-}
-
-function setDeleteCard() {
-
-}
-
-function setBanCard(ban, email) {
-
 }
